Updates database and print queue UI
Updates the database schema to include `additional_parts_cost` and `list_price` to the `products` table, and creates a `product_filaments` junction table.

Adds a `position` column to the `print_queue` table for ordering, and creates an index for faster ordering.

Refactors the print queue UI for better printer selection and status updates.
Updates settings component to improve data handling and UI. Includes basic sanitization of company names for display and edit.